requirements.txt
sqlite-utils>=2.16
requests>=2.24.0
selenium>=3.141.0
PyYAML>=5.3.1